requirements.txt
asgiref==3.4.1
dj-database-url==0.5.0
Django==4.0.1
django-cors-headers==3.10.1
django-heroku==0.0.0
django-postgres-extensions==0.9.3
djangorestframework==3.13.1
djangorestframework-simplejwt==5.0.0
djongo==1.3.6
dnspython==2.1.0
numpy==1.22.3
pandas==1.4.1
Pillow==9.0.0
psycopg2-binary==2.9.3
PyJWT==2.3.0
pymongo==3.12.1
python-dateutil==2.8.2
pytz==2021.3
reportlab==3.6.6
six==1.16.0
sqlparse==0.2.4
whitenoise==5.3.0
